requirements.txt
# MeetSpot Dependencies
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-core==2.33.2
httpx==0.28.1
aiohttp==3.12.15
aiofiles==24.1.0
jinja2==3.1.6
python-multipart==0.0.20
loguru==0.7.3
tomli==2.1.0
python-dateutil==2.9.0
jieba==0.42.1
whitenoise==6.6.0
slowapi==0.1.9
markdown2==2.4.12
asgiref==3.8.1
requests==2.32.3
beautifulsoup4==4.12.3
sqlalchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
websockets==12.0
